refactor(files): emojis, badges, constants
- Updated Database Schema, and reformatData

- Removed util/flips
- Removed util/emojis
=> Merged into util/constants.js file (with new file prefix)

- Updated file pathways in other files
