+++ src/commands/information/userinfo.js
@@ -1,6 +1,6 @@
 /* Import required modules and files */
 const { SlashCommandBuilder, EmbedBuilder } = require('discord.js');
-const emojis = require('./../../util/emojis');
+const { emojis } = require('./../../util/constants.js');
 const { database } = require('./../../util/functions.js');
 
 module.exports = {
