--- src/util/emojis.js
@@ -1,56 +1,0 @@
-module.exports = {
-
-	order: [
-		'dev',
-		'partnered_dev',
-		'support',
-		'flip',
-		'flip_plus',
-		'minigame',
-		'minigame_plus',
-		'register',
-		'collector',
-		'collector_plus',
-		'rich',
-		'rich_plus',
-		'niceness',
-		'bughunter',
-		'bughunter_plus',
-	],
-
-	blackbelt: '<:belt_black:975078485737500692>',
-
-	false: '<:false:956523937691500616>',
-	true: '<:true:956523933522362482>',
-
-
-	dev: '<:badge_developer:956323651941183488>',
-	partnered_dev: '<:badge_partner:956323653564370966>',
-
-	support: '<:badge_support:956323525105446942>',
-	register: '<:badge_registered:956323523352231946>',
-	niceness: '<:badge_niceness:975078534554992720>',
-
-	flip: '<:badge_flipper:956323329688633354>',
-	flip_plus: '<:badge_fliptician:956323335774552084>',
-
-	collector: '<:badge_collector:956323383103078440>',
-	collector_plus: '<:badge_scavenger:956323385762263100>',
-
-	minigame: '<:badge_minigamer:956323434265182288>',
-	minigame_plus: '<:badge_pro_gamer:956323436286836826>',
-
-	rich: '<:badge_rich:956323487847440425>',
-	rich_plus: '<:badge_wealthy:956323490137505872> ',
-
-	bughunter: '<:badge_bug_hunter:956323559544872981>',
-	bughunter_plus: '<:badge_bug_poacher:956323564766757015>',
-
-
-	gold_tier: '<:badge_cf_gold:956323628817997824>',
-	coin_gold_tier: '<:coin_goldtier:832295667795624027>',
-
-	platinum_tier: '<:badge_cf_platinum:956323631397503037>',
-	coin_platinum_tier: '<:coin_platinumtier:832295735445553152>',
-
-};
